% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/awsjavasdk.R
\name{aws_sdk_present}
\alias{aws_sdk_present}
\title{Confirm presence of core SDK jar}
\usage{
aws_sdk_present(assert = FALSE)
}
\arguments{
\item{assert}{If TRUE, the absence of the SDK will throw an error}
}
\value{
Boolean reflecting whether the SDK could be located
}
\description{
Confirm presence of core SDK jar
}
\details{
Note that this only checks for the ability to locate the core AWS SDK jar
}
\examples{
aws_sdk_present()
}

