% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/awsjavasdk.R
\name{aws_sdk_version}
\alias{aws_sdk_version}
\title{Find AWS SDK Version}
\usage{
aws_sdk_version()
}
\value{
character, AWS SDK Version and standard character prefix \code{aws-java-sdk-} as in the unpacked directory name
}
\description{
Find AWS SDK Version
}
\examples{
\dontrun{
aws_sdk_version()
}
}
