% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/awsjavasdk.R
\name{install_aws_sdk}
\alias{install_aws_sdk}
\title{Install the AWS Java SDK}
\usage{
install_aws_sdk(root = .package_envir_get("sdk_file_root"))
}
\arguments{
\item{root}{The root folder that will contain the SDK}
}
\value{
boolean; TRUE if the SDK can now be located, FALSE if it can not be located
}
\description{
This function fetches and unpacks the AWS Java SDK to a package managed directory.  
If the Java SDK is already present, 
  it will wipe that location and re-download and unpack the
  file present at https://sdk-for-java.amazonwebservices.com/latest/aws-java-sdk.zip
}
\examples{
\dontrun{
install_aws_sdk()
}
}

