% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/awsjavasdk.R
\name{set_sdk_file_root}
\alias{set_sdk_file_root}
\title{Specify an alternate root directory in which to store and retreieve the SDK.}
\usage{
set_sdk_file_root(root = .app_dir$data(), confirm = FALSE)
}
\arguments{
\item{root}{character a path location the contents of which may be deleted without confirmation}

\item{confirm}{boolean Allows the use of a non-empty sdk_file_root}
}
\description{
Specify an alternate root directory in which to store and retreieve the SDK.
}
\note{
This feature is required by the CRAN maintainers.  Although it is useful, it has not been as thoroughly tested as using the default package selected directory.
}

